feat: Full Spec Compliance — 9-Layer Pipeline, Session Rules, Quality Scoring & Dashboard (#14)
* test: add 26 dashboard GUI responsiveness tests

Every frontend fetch() call in index.html is now covered:
- 17 GET endpoints (signals, status, market, backtest, news, logs, feargreed,
  dex pools, data sources, auto status, config, trades, indicators, orderbook,
  candles, realtime snapshot)
- 7 POST action endpoints (trade, close-all, breakeven, auto toggle, mode
  toggle, config save, config test)
- SSE stream route registration check
- Full fetch-target inventory validation

Fix: SSE streaming test no longer hangs (route check instead of body read)

216 tests pass (26 new GUI responsiveness tests)

* fix: add rate limiting to all order/cancel calls and fix POSITION_CLOSED kill switch format

* feat: iterations 2-5 — exchange-side SL/TP, hedge mode, tick rounding, production hardening

## Exchange-Side SL/TP (Critical)
- STOP_MARKET + TAKE_PROFIT_MARKET placed on exchange after every fill
- reduceOnly=true in one-way mode; positionSide=LONG/SHORT in hedge mode
- Auto-detect hedge mode via fapiPrivateGetPositionSideDual()
- Tick-size and step-size rounding from CCXT market info (floor rounding)
- OCO: SL fill cancels TP, TP fill cancels SL
- Crash-safe: exchange-side orders survive bot termination
- Reconciliation re-registers existing SL/TP on restart

## Production Hardening (Iterations 2-4)
- Kill switch directly cancels exchange orders via executors
- Circuit breaker trips on SL placement failure
- WAL-based trade persistence for crash recovery
- Startup reconciliation rebuilds state from exchange truth
- Rate limiting on all order/cancel API calls
- ML model SHA-256 integrity verification before loading
- Idempotent order placement prevents duplicate fills

## Tests
- 20 new exchange-side SL/TP integration tests
- 19 iteration-4 proof tests (8 non-negotiable requirements)
- Full suite: 287 passed, 0 failed

* feat: ARMS-V2.1 gap closure, paper auto-trading pipeline, UI responsiveness

- Close 23 remaining ARMS-V2.1 spec gaps (TWAP/Iceberg splitter, ShadowSL,
  strategy modules, pair registry, safe mode, audit trail, state recovery)
- Wire end-to-end paper trading pipeline: PaperFeed → EventBus → DataManager
  → SignalGenerator → PaperExecutor with /api/paper/trades endpoint
- Fix standalone dashboard to use real EventBus (was DummyEventBus)
- Paper mode tuning: redistribute weights (tech 50%, ML 40%), lower min_score
  to 0.03, disable HTF confirmation for limited data feeds
- Fix UI: remove duplicate HTML document, add collapsible sidebar sections,
  responsive layout (<900px drawer mode), chart ResizeObserver
- Fix orderbook display (depth=5, both asks+bids visible)
- Add 879+ tests covering all new modules and fixes
- Add Alembic migrations for PostgreSQL audit trail schema
- Add alert_manager, monitoring metrics enhancements

* fix: resolve CI failures — portable paths, pandas-safe Keltner comparison

- Replace hardcoded /workspaces/nueral-trader-5/ paths with
  Path(__file__).resolve().parent.parent.parent for CI portability
- Fix Keltner channels test: use aligned mask instead of dropna() which
  can produce misaligned Series on different pandas versions

* feat: full spec compliance — 9-layer pipeline, session rules, quality scoring, risk guardrails, dashboard panels

## Signal Pipeline (§3-§6)
- Add SignalType enum (A/B/C/D/COMPOSITE) with priority ordering
- Implement 6 trading session rules + ICT killzone enforcement + weekend halt
- Add 0-100 quality scoring with 8 weighted components (min 65 gate, 80+ boost)
- Replace binary HTF check with weighted MTF scoring (3TF/2TF/1TF position sizing)
- Add signal type classification based on SMC state (FVG/OB/liquidity/breakout)

## Regime & Risk (§7-§9)
- Add regime-specific risk percentages (trending=2%, ranging=1.5%, breakout=2.5%)
- Implement regime transition protection (50% size, 0.5x SL, 2-candle window)
- Fix daily loss halt: 1h → 24h per §9
- Fix weekly DD limit: 6% → 8% per §9
- Fix max exposure per symbol: 10% → 6% per §9

## Risk Guardrails (§9, §13)
- Add flash crash detection (5% drop in 5min window)
- Add consecutive loss circuit breaker (3-loss halt for 1h)
- Implement half-Kelly Criterion position sizing (rolling 100-trade window)

## Technical Indicators (§3)
- Add Donchian Channels (20-period)
- Add Vortex Indicator (14-period)
- Add Awesome Oscillator (5/34 SMA)

## Frontend
- Add 5 new API endpoints: /api/layers, /api/session, /api/quality, /api/regime, /api/guardrails
- Add Pipeline tab with 9-layer confirmation grid + quality score breakdown
- Add sidebar panels: Session Clock, Regime Status, Guardrails

## Tests
- 879 passed, 0 failures
- Fix 4 test callsites for updated HTF return signature (2→4 tuple)

+++ execution/okx_executor.py
@@ -235,7 +235,9 @@ class OKXExecutor:
             return None
 
         if abs(funding.funding_rate) > funding_threshold:
-            direction = "long" if funding.funding_rate > 0 else "short"
+            # positive funding_rate → longs pay → go short to receive funding
+            # negative funding_rate → shorts pay → go long to receive funding
+            direction = "short" if funding.funding_rate > 0 else "long"
 
             return {
                 "symbol": symbol,
